added Linked tags
added <Link> tags  for redirecting the URL path upon the click events in 'movie-card.jsx' and 'movie-view.jsx' files.

since the <Link> tag are used for handling the click events, the props that we used for this purpose were removed, in 'movie-card.jsx' and 'movie-view.jsx' files.

comments were added throughout the files