added navigation bar with react routing
'navigation-bar.jsx' file is created to using React Bootstrap, and then linked to the routing in 'main-view.jsx'.

In 'main-view.jsx'  file, dynamically added the logic for logging-out as a function of a variable- to be passed into as a prop for the <Link>  tags in the 'navigation-bar.jsx' file.