isolating the problem in <ProfileView> rendering
simplified the <ProfileView/> code to  isolate the problem, in 'profile-view.jsx' file.

commented out the code for the project line 48-113, and practice fetching using a simple return() at line 16-42 in 'profile-view.jsx' file

changed the prop into a json object so the username property can be accessed.  line 11-13 in 'profile-view.jsx' file.