added code for the set_favourite feature
In 'main-view.jsx'  file:
the handler function for setting a movie as a favourite is passed into the child components on line 256-259 and 291-294

In 'movie-card.jsx' file
helper function (line 10) added with useEffect (on line 32)  to keep the state synchronised between the components.

logic for handling the click events of the favourite buttons on line 17-43 and line25-40.

buttons the 'favourite' functionality on line 57-71

on 'movie-view.jsx':
button functionality added on line68-75

the same 'favourite' functionality for the handlers in 'movie-card.jsx' file, including the helper function,  also added in this file on line7-42