refactor: shifting the loading state from global to dynamic
In 'favouriteSlice.js':
the loading state was changed  (line 5) from a global string to a dynamic object with each movie's movieId serves as a key.
this allows for the application to track the loading state individually for each movie.

the extraReducers in the slice is updating the loading state dynamically line 90-112.
When the addFavouriteMovies or removeFavouriteMovies thunk is dispatched, the pending case sets the specific movie's loading state to 'pending'.
Upon fulfilled or rejected, the specific movie's key in the loading object is updated to 'succeeded' or 'failed', respectively.

New selector function introduced to retrieve the loading atate for a specific movie by its movieId- exported out the favouriteSlice.js file on line 128.  used by the components to access and respond to the loading state-  in 'movie-card.jsx' and 'movie-view.jsx' on line 15 and 19, respectively.

In 'movie-card.jsx' file:
conditional rendering of the buttons now use the loading state specific to the movie being interacted with. On line 54-56 and 61-65.

these changes ensures that only the movie being updated reflects a loading state, preventing all buttons in the UI from being affected during an API request- providing accurate feedback to the user.